Refactor config-channels component to be useable as states picker for recurring states
